remove Default derives for instance spec types
This frees the spec types from having to derive `Default`. This is not
so bad for many types, but (AFAICT) there's no way to patch
`#[derive(Default)]` onto an enum type, because there's no way to tell
Progenitor which variant is the default.

In any case, users of the client lib already have access to an instance
spec builder, so the value of creating a default spec is diminished (and
anyone who really wants to build a spec by hand can still do that, they
just have to choose their own default values).