[WIP] use SinkBuffer instead of sending characters one by one
This is needed to prevent flakiness in PHD tests that send multi-line
commands (like the migration tests that type a large shell script into
the terminal).

When a Linux guest shell echoes a multi-line command, it writes out a
`CR LF > \u{20}` sequence (i.e. it breaks the line, then writes a `>`
and a space) at the end of each line. It appears that if the active
terminal receives a character while it's writing this sequence, the
character will be echoed immediately, even if there are other characters
that were buffered ahead of it that haven't been echoed yet. Extra
characters buffered this way are still echoed in the correct place in
the buffered-character sequence.

Brokering UART writes through a `SinkBuffer` drastically reduces the
chance this problem will occur: when the buffer is in use, most writes
to the UART happen in batches when the UART calls the buffer's "ready to
write" notification routine. Batching writes this way reduces the chance
that a write (to the guest) will be interleaved with a guest line break
sequence in a way that produces extra characters.